Bugfixes & Cleanup (#120)
* Bugfixes & Cleanup
* Cleaned up GameMods.hpp & fixed all macro toggles
* Made world border unpickable/unhighlightable
* Lowered mob reach range for creative mode from 32 blocks to 7 (as in 0.12.1)
* Fixed uninitialized variables (bugs could appear depending on the compiler)
    * "PathfinderMob::m_pAttackTarget" (caused a crash)
    * "Entity::field_C4" (caused entities to suffocate always)
* Fixed Random.cpp for compilation with C++03 by using "std::numeric_limits" instead of "INFINITY"
* Documented Texture::field_C as m_hasAlpha
* Deleted unused "GLTexture" struct
* Removed ModelPart::renderHorrible
* Added other mob textures to .gitignore

* Fixed incomplete "Texture::field_C" to "Texture::m_hasAlpha" rename

* Bugfixes for Windows SDL2 Visual Studio compilation
* It still does not work. I'm not even sure how it broke in the first place.

* Added Panorama Toggle
* Added option & enhancement macro for enabling/disabling the panorama menu background

* Documentation & Decompilation
* Fixed a bug which caused mobs to not catch fire
* Added Flint, Porkchops, and Eggs
* Mobs can now drop loot upon death
* Added Chickens
* Added scaffolding for gamemodes in replication and level data. This can be enabled with 'TEST_GAMEMODE_REPLICATION'.
* Documented a number of variables

* Partial Network Protocol v3 Support

* Documentation & More Backporting

* Minor Additions
* Add EntityType
* Add MobFactory
* Add PacketUtil
* Remove TLong/LongHack remnants
* Remove old minecraftcpp.vcxproj file

* Fix typo in CMakeLists.txt

* Survival Mode Bugfixes
* Reverted protocol version back to 2

* Cleanup
* Added Cow
* Made tons of stuff const
* Fixed-ish pathfinding

* Fix un-renamed Texture.m_hasAlpha that loved to hang around for some reason

* Fixed minor bugs that gave compile errors :(

* Updated .gitignore for 0.7.0 assets

* * Fixed cloud scroll speed
* Replaced global variable with LevelRenderer::areCloudsAvailable()

* macOS & iOS Bugfixes

* * Added /summon command
* Fixed crashing from moving infinite distane
* Minor aesthetic tweaks to DirectConnectScreen
* Cleanup some, but not all, VS project files

* Bugfixes & Cleanup
* Added missing mob .cpp files to CMakeLists.txt
* Cleaned up Visual Studio project files
* Fixed SDL2 client building for Windows on VS
* Fixed mouse centering on SDL2 clients

* Standardized coordinate structs
* Added ChunkPos
* Added Vec2
* Added Facing::Name enum for cardinal directions
* Added TilePos (this is a must-see!)
* Added ChunkTilePos

Cleaned up input handling
* Added SDL2 controller support
* Added read-only "Use Controller" "option" to visualize current input handler
* Created ControllerBuildInput & ControllerMoveInput
* Moved mouse-specific build-input-handling code from Minecraft class to new MouseBuildInput class
* Refactored BuildActionIntention

Cleaned up VS project files
* Fixed VS include path properties
* Corrected SDL2 client building in VS

Other
* Moved iOS-specific assets to accurate location
* Abstracted EmptyLevelChunk CPP into its own .cpp file

* * Fixed various PR issues
* Added icon for Win32 client
* Enabled UI button tab navigation for gamepad
* Fixed a couple of incorrect gamepad input mappings
* Mitigated LocalPlayer::isImmobile() bug to gamepad only
* Tweaked ControllerTurnInput to closer match Legacy Console Edition's turn behavior
* Documented various fields
* Made timing more percise for Windows

* Fixed instance of VS rename feature being braindamaged

* * Removed Facing.cpp

* Minor fixes in Entity and ControllerMoveInput

* Fixed Mob travel bug

* Remove Facing.cpp from CMakeLists.txt

* Fix Level::getCubes ambiguous TilePos constructor call

* Fixed various bugs & building for macOS + iOS

* Minor compatibility bugfixes

* Fixed Xcode project settings

* Additional nullptr #ifndef check

* Documentation + Bug Fixes
* Fixed networking bugs

* Fixed PR stuff

* Visual Improvements
* Added normal lighting support
* Added the Sun and the Moon
* Added entity circle shadows
* Added vignette
* Added manual mob rendering offset to fix floating mobs

* Xcode & Mobile Bugfixes
* Fixed glColor3f calls for GLES
* Added texture availability check to AppPlatform_iOS
* Fixed building via Xcode

* Minor Fixes
* Set iOS deployment target from 6.0 back to 3.0
* Set default GameMode from Survival back to Creative

* Update CMakeLists.txt

* Update TextInputBox.hpp

* Update Screen.cpp

* Update TextInputBox.hpp

* Update TextInputBox.cpp

* Update TextInputBox.cpp

* Update TextInputBox.cpp

* Fixed bug which caused a segfault on Linux
- Removed extra glPopMatrix call in ItemInHandRenderer::render()

* Sheep + SynchedEntityData
- Added Sheep
- Added a mostly-finished decomp & revamp of SynchedEntityData
- Fixed MobRenderer scaling issue which caused all Mobs to render slightly too small & in the ground

* Moved SynchedEntityData DataType mappings from .hpp to .cpp

* Update SheepRenderer.cpp

* Documentation & Bugfixes
 - Fixed minor pathfinding issues, still more to fix.
 - Fixed additional fire rendering for entities.
 - Fixed footsteps being disabled for all entities but the player.

* Minor Cleanup

* Fixed Pathfinding Bugs
- Fixed issue where mobs would think they're stuck when in water
- Documented more Node & PathFinder fields

* Fixed Building for VS2010
- Also fixed initialization errors in HumanoidModel

* EntityTypeDescriptor & EntityTypes
 - Added EntityTypeDescriptor
 - Added EntityType class
 - Added EntityCategories class
 - Fixed unecessary game crash that would occur when OpenAL throws an error

* ISO C++ Compliance

* Fixed MobFactory macro

* Xcode Fixes & C++03 Compatibility

* Update build.yml

* Homebrew hack

* Homebrew hack

* Homebrew hack

* Minor fixes

* Revert "Update build.yml"

This reverts commit 7e65e63dd93583b65b45bb3ccdda46c0c0ec7e55.

* Hopefully fixes GitHub Actions CMake

---------

+++ source/world/level/levelgen/feature/LargeCaveFeature.cpp
@@ -42,7 +42,7 @@ void LargeCaveFeature::addTunnel(int x, int z, TileID* tiles, float rx, float ry
 
     float yRota = 0.0;
     float xRota = 0.0;
-    Random random(m_random.nextLong());
+    Random random(m_random.nextInt());
 
     if (dist <= 0)
     {
